+++ packages/meteor-tool/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: 'The Meteor command-line tool',
-  version: '2.7.0-rc.0',
+  version: '2.7.0-rc.3',
 });
 
 Package.includeTool();