+++ packages/react-fast-refresh/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'react-fast-refresh',
-  version: '0.2.3-rc270.0',
+  version: '0.2.3-rc270.3',
   summary: 'Automatically update React components with HMR',
   documentation: 'README.md',
   devOnly: true,